Introduce survival curves conditional on parity progresion
Instead of using standard survival curves which cannot be
easily compared when the likelihood of having a next child
differ across groups the new graphs show survival conditional
on progressing to next birth.

Also tried to run combined code across periods. Seems to wash out
some of the differences across periods and it might therefore
be better to keep the standard way of running the regressions.