Rename all graphs and update Makefile with dependencies/targets
Instead of a numbering system, which was hard to read, all graphs
are now named with urban/rural and the sex composition of prior
children, if any.

All references in paper and appendix changed to fit the new
naming system.

The Makefile was update as well and now include the relevant
targets and dependencies with the new naming system. In principle,
it should be possible to shorten the generation of the dependencies
and targets, but that would require an if statement and the if
function in Make runs of whether a given piece of text is missing
or not, rather than a standard condition.