Clean up and fix typos in targets and dependencies
Targets were generated without the sex composition for children
because of missing parentheses and the name of the target for
spell 3 was wrong.

Added phony "results" to avoid having to compile whole paper each
time checking results.

Clean up text and add caveat to introduction of Makefile for
the complications.