Initial version of bootstrap standard errors for median spell lengths
To get bootstrapped standard errors for median spell lengths (and
other statistics) required created a new program (bootspell.do) that
the bootstrap command can call. This is required because the resampling
needs to take place on the original sample, rather than the sample
after it has been expanded to number of periods observed. The
an_bootstrap_table.do file will eventually create the table, although
currently all it does is to loop over spells and periods and call
bootspell.do for each combination and print the results to the
screen.

bootspell.do has two parts. The first expands to number of periods,
call the baseline hazards, creates the non-proportional variables,
and runs the estimation. Nothing is saving here. The second uses
the estimation sample to predict median spell length by individual
and returns the average median spell length by group (area and
sex composition). It is also set up for eventually returning
percentage boys, but does not do so currently.

The naming convention of explanatory variables was not completely
consistent across the second and subsequent spells for sex composition,
so spell 2 will not run at the moment (spells 1, 3, and 4 all work).

Baseline hazard are in a separate folder. The idea is one file per
education level, although currently only high education are in there.
In each file are 12 programs (4 spells by 3 periods). These programs
can be called as long as the main hazard files are loaded. This was
a bear to get to work. Stata does not look including files that
has loops in them from another program, which is why these are
now individual programs that each return the number of piece-wise
linear hazard sections for use. Hence the use of "loc i = `r(numPer)'"
after each call to the hazard program.