Split bootstrap into calculation and table creationg and fix issues
Main fix: bootspell.do had an "eststo" on the regression command, which
lead Stata to reach its system limits at about 300 regressions. Removing
that solves the problem causing bootstrap to fail at what appeared
different education/spell/period combinations.

I split the bootstrapping and the table generation into two files
because the bootstrapping now save the results to a dta file
rather than a matrix. This allows for easy changes to the table
without having to run the entire bootstrapping calculation again.
This becomes especially important with large number of repetitions
for the bootstrap.

The table generation still creates matrices to ease the table writing,
but now creates these before starting on the table itself. The
creation is fast enough to not create any problems. One downside
of using saved data is that it is necessary to clear results, simply
loading the bootstrap result data with clear and then calling "bstat"
does not get rid of the old results.