Fix birth prob; added iweights to summarize and new statistics
Original version of probability of birth was actually probability of
no birth! Added "1 -" to fix.

Calculation of statistics now uses iweight based on the probability of
a birth by the end of the spell. According to the manual this should
produce the correct weighting despite iweight generally not recommended.

Added probability of a birth by end of spell as a new statistics. This
is not weighted since I am here interested in the standard average
across women in the sample.