Initial version of combined samples bootstrap for testing dif-in-dif
Different period samples are combined and run with full period interactions
to allow testing of differences in statistics across periods and
"difference-in-difference" of changes in duration statistics.

This does not use baseline hazards but rather the individual
3-months periods because no graphs are produced.

This is implemented as an ado, but the periods and spell covered are
hard-coded into the file.

Two parts are missing:
- allow call for periods and spell
- combine durations if multinomial logit is not concave.