Some region bootstrap analyses do not converge, so limit to high education
The problem is mainly for the medium education women, where there are a low number
of observations, especially for the southern parts of India. Since high education
women are more likely to use sex selection focus is now on them.

Also has the advantage of cutting down on run time.